feat(CommentTeaser): cleanup
